split process
